[payment] Fix race condition on multiple new Team Plan joins
